+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.2.0 - 2024-10-08
 
+* NEW FEATURE- added `cancel` command to allow cancellation of in progress jobs.
 * Updated/migrated clap to v4.4.
 * Updated dependency quinn-proto to latest to fix security vulnerability.
 * BUGFIX- the alias "sv" was assigned to both "submit-vis" and "submit-volt". "st" has now been assigned for "submit-volt" to avoid the duplication.